\name{FERG2-package}
\alias{FERG2-package}
\alias{FERG2}

\docType{package}

\title{Data and Functions for the FERG2 Computational Framework}

\description{
The \pkg{FERG2} package provides a series of data inputs and helper functions for the FERG2 computational framework.
}

\details{
  \tabular{ll}{
    Package:    \tab  FERG2\cr
    Type:       \tab  Package\cr
    Version:    \tab  0.0.3\cr
    Date:       \tab  2025-04-07\cr
    Authors:    \tab  Brecht Devleesschauwer\cr
    Maintainer: \tab  Brecht Devleesschauwer <\email{brechtdv@gmail.com}>\cr
    Depends:    \tab  R (>= 4.4.0)\cr
    License:    \tab  GPL (3)\cr
}

Available functions in the \pkg{FERG2} package:


Available datasets in the \pkg{FERG2} package:
  \tabular{ll}{
    \code{\link{countries}} \tab WHO member states: names, ISO3 codes, ISO2 codes, WHO regions, FERG2 subregions, FERG1 subregions.\cr
  }
}

\keyword{package}
